backend/pyproject.toml
[project]
name = "ai-virtual-interior-designer"
version = "0.1.0"
description = ""
authors = [
    {name = "Narender Keswani",email = "narender.rk10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-genai (>=1.11.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
